requirements.txt
polars
pandas
geopandas
shapely
numpy
matplotlib
cartopy
streamlit
requests
python-dotenv
duckdb
quixstreams
